Remove menu items instead of disabling them
When the size of menu is huge (hundreds of menu items), disabling
particular menu items in a loop is very slow. Profiling the code shows
that a lot of time is spent calculating menu width based on a different
set of visible menu items. This is happening somewhere in NSMenu
internals and it seems like cannot be controlled manually.

In order to avoid slow computations, we now simply remove menu items
that don't match search filter and it works fast enough on hundreds of
records.

Closes #13