Refactor menu loading to avoid rebuilding its items every time
Previously, when Maccy menu opens, it would load all the history from
storage defaults and rebuild the menu. Now, it's only build on launch
and then it's constantly being rebuilt in the background as soon as new
item is copied to the clipboard or removed from it.

Hopefully it fixes #36.