Respect search when selecting via shortcuts
This fixes a regression that resulted in search being ignored when user
selected an item using keyboard shortcut. The bug turned out to be super
simple even though I've spent hours debugging it. The issue was
essentially the following actions:

1. User searches something.
2. Key equivalents are updated correctly.
3. User presses key equivalent.
4. Window starts closing.
5. FilterMenuItemView clears the search query.
6. Key equivalents are reset.

Because of the last action, the original (unsearched) item was selected.
The fix was obvious - do not clear search query when window closes but
rather when it's opened.

Closes #67.