Add custom menu title color to blend better in high contrast mode
In normal contrast mode the background colors of NSMenuItem with custom
view are similar to other menu items. But in high contrast mode the
colors are slightly lighter/darker than they should be. To workaround
the problem, we're using custom-tailored colors for high contrast mode.
For normal one they should not be altered at all not to interfer with
transaprency in default UI (however, the colors should still be
specified for asset so I simply used purple).

See https://github.com/p0deje/Maccy/issues/64#issuecomment-570768325
for more details and screenshots in high contrast mode.