Fix search to find all items with the same menu item title
The old version of code would fail when we have two IndexedItem with the
same `value` but different `item`. In this case, it would erroneously
pick only the first one. It would also make key equivalents be ordered
incorrectly.

An example of such data is copying two identical Github comments - they
produce the same `.string` pasteboard data, but different `.html`. They
should be shown as 2 separate menu items and should be properly found.