Do not attempt to move pinned items when they are copied
Current behavior has a bug which only happens when pinned items are
shown at the bottom and visible menu items is limited. In this case,
selecting pinned item causes a crash because it attempts to insert its
menu items at the index that is beyond visible menu items (e.g. it
should insert at 999 because it's the last item but maxMenuItems is set
to 20).

Another solution could be to avoid removing items when reaching
maxMenuItems in favor of setting isHidden to true, but that requires
performance evaluation first.

In any event, there is not much point in moving pinned items around.