Bump minimum supported version to macOS 10.14 Mojave
Since v0.15.0 the sandbox has been enabled in the app. This in fact
broke pasting on High Sierra. Unfortunately, the sandbox prevents
emitted paste event from reaching the application on this particular
version of macOS. Apple Developer support stated that this is a bug
which has probably been fixed in Mojave, but won't even be in High
Sierra.

At the same time, I haven't found any reliable alternative to pasting
event. One of those was to use AppleScript but it suffers from other
problems. See discussion in #161 for more details.

In the very end, it seems like there are 2 ways: either stop supporting
High Sierra as pasting doesn't work there or disable the sandbox. Given
I plan to distribute the app in the App Store, the former is really the
only way to go for now.