Ensure focus is returned after accessibility permissions modal
When Maccy asks for accessibility permissions on paste, it disable focus
returning to previous window because the accessibility modal should be
kept focused. However, after user closes the modal, we should make sure
that Maccy returns focus to the previous window on the next paste. It
was not before this commit.

Steps to reproduce:
1. Open Maccy with no accessibility permissions.
2. Select item to paste.
3. Select "Deny".
4. Open Maccy.
5. Select item to copy.

Ensure that the previous window gets focus back.