Workaround Korean input handling issues (#476)
Whenever we call becomeFirstResponder() on the search field, it breaks
Korean input handling where typing "rkadnf" should end up with "강물".

Ideally we'd just call some method to re-evaluate the string in the
search field and replace with proper characters, but I have not found a
way to do that yet.