Replace simple tooltips with a rich preview popover
This commit implements a long-wanted feature where the contents of each
history item is shown in a popover with higher-resolution images and
mulitline text. It's implemented by creating an extra NSMenuItem of 0.1
pixels height, hidden and implementing a custom view. Then once the
history item is highlight in the menu, an NSPopover is shown next to
this hidden menu item. The contents of the popover as loaded from
Preview.xib. The delay until popover is shown attempts to mimic default
tooltip behavior, with about 1.5 seconds delay for the first popover and
0.1 seconds delay for subsequent popovers. The delay can also now be
changed based on the settings, though it's not yet implemented.

Fixes #267