Fix Sonoma Beta 2 menubar hang/crash
This fixes the hang and crash that occurs since macOS Sonoma Beta 2.

Maccy used to hide the "Maccy" title and search bar within the menu by setting the height of the NSView to a small value.

In Beta 2, modifying the frame in any way results in a crash.

While this is likely an OS bug, since setting the height to 1 is a hack anyway, and the header only contains the title and search bar, not adding the header when both are disabled seems like a cleaner solution to me.