Avoid menu from moving upwards while scrolling
Until now if the menu has to show a scrollbar it will resize upwards while scrolling downwards.
To prevent this we use the confinementRect(for:on:) method of NSMenuDelegate to limit the size
of the popup. To avoid computing the correct bounds ourself we show the popup and immediately hide
it while saving the bounds from the window.