Ensure fuzzy search sort order is preserved
This requires to maintain two separate references to stored items (`all` and `items`). The former is what is given by the storage + sorted by an algorithm specified in settings. The latter is what is shown in the popup with elements being hidden and re-ordered during search.

This is a bit more work and I'd prefer to keep using a single list and rely on `.isVisible`, but that means we'll ned to resort items after every fuzzy search.